feat: Fetching ICE config from Catalyst Server
